+++ Packs/Expanse/Integrations/Expanse/CHANGELOG.md
@@ -3,8 +3,8 @@
   - Fixed an issue related to the ***ip*** command where an error is generated if the API returns a partial response.
   - Added friendly values for various empty fields returned by the ***domain*** command.
   - Added Support for pulling behavior data to create new incidents.
-  - Added Support for the **behavior** command.
-  - Added Support for the **certificate** command.
+  - Added Support for the **expanse-get-behavior** command.
+  - Added Support for the **expanse-get-certificate** command.
 
 ## [20.3.3] - 2020-03-18
   - Updated the Authorization header for the Events API to use the correct token.